Fix problems with build from source distribution
Building from source continud to be problematic. The culprit turned out to be
that the generated C/API files _irimodule.c and _iriplasmodule.c were treated
as source files, whereas they should be treated as generated files. Doing an
in-place build didn't help it that respect.

But the real issue was, that the files iri.pyf and iriplas.pyf were _not_ part
or the source distribution. This makes sense, because they should ideally be
generated by f2py from the Fortran sources. However, since those files had to
be tweaked a bit, they _must_ be considered as source files now. Once they were
added to the source distribution, all issues were magically resolved. Even the
trick with the overridden `build_src` class is no longer necessary.