Fix pkg_resources issue (#61)
* Add more basic tests

Added a few more basic tests that simply try to instantiate a few classes and invoke a method on the instances.

* Regenerate SWIG wrapper

The old SWIG wrapper caused issues with Python 3.12. Regenerated the `EMM_Model` bindings with the latest SWIG.
TODO: These bindings should be generated during the build, and not be put under version control in Git.

* Replace obsolete pkg_resources with importlib.resources

Replaced obsolete `pkg_resources` with `importlib.resources` or its back-ported cousin `importlib_resources` for older Python versions.